--- line.py
@@ -1,12 +1,0 @@
-from point import Point
-
-
-# classe que define uma linha basica dentro do universo de representacao
-class Line():
-    # construtor
-    def __init__(self, start: Point, end: Point):
-        self.start = start
-        self.end = end
-
-
-# end of class Line
