--- point.py
@@ -1,10 +1,0 @@
-# classe que define um ponto basico no universo de representacao
-class Point():
-    # double x_, y_;
-    # construtor
-    def __init__(self, x: float, y: float):
-        self.x = x
-        self.y = y
-
-
-# end of class Point
