--- polygon.py
@@ -1,14 +1,0 @@
-# classe que representa um poligono generico
-
-
-class Polygon():
-
-    # list points_;
-
-    # construtor
-    def __init__(self, points):
-        self.points_ = points
-
-
-
-# end of class Polygon
